improvements after comments
