+++ README.md
@@ -55,3 +55,10 @@ import (
 	_ "github.com/redhat-appstudio/e2e-tests/tests/has"
 )
 ```
+
+# Debugging tests
+## In vscode
+There is launch configuration in `.vscode/launch.json` called `Launch demo suites`. 
+Running this configuration, you'll be asked for github token and then e2e-demos suite will run with default configuration.
+
+If you want to run/debug different suite, change `-ginkgo.focus` parameter in `.vscode/launch.json`.
